Wave animation
The wave consists of wavefronts. The crests are colored red whereas the troughs are blue. To show the curvature of crests and troughs, the opacity of the color varies slightly from the start of a crest/trough to the end of it. The wave decreases its intensity, which is shown as the opacity of colors, as the square of distance after traveling 100px and gets completely destroyed after traveling 500px.